[DSCP-483] Finished splitting Dscp.Core.Foundation.Educator
[DSCP-483] Some more grouping
