+++ core/src/Dscp/Core/Foundation/Educator/ATG.hs
@@ -1,0 +1,37 @@
+
+module Dscp.Core.Foundation.Educator.ATG where
+
+import Control.Lens (makeLenses)
+
+import Dscp.Core.Foundation.Educator.Subject
+import Dscp.Util
+
+{-
+
+@flyingleafe: I implemented a version of ATG without "etc" vertices
+because I don't quite see the reason for having them.
+
+-}
+
+-- | Activity Type Graph node. Implementation without "etc" vertices.
+-- TODO: should we use 'Vector' for more efficient indexing? or we don't
+-- care, because it should be in DB somehow anyway?
+data ATGNode = ATGNode
+    { _atgnSubjectId :: !(Id Subject)
+    , _atgnChildren  :: ![ATGEdge]
+    } deriving (Show, Eq, Generic)
+
+-- | Edge pointing to the children node in ATG.
+data ATGEdge = ATGEdge
+    { _atgeWeight :: !Float
+    , _atgeChild  :: !ATGNode
+    } deriving (Show, Eq, Generic)
+
+-- | Activity Type Graph itself is just a list of root nodes.
+newtype ATG = ATG
+    { getATGRoots :: [ATGNode]
+    } deriving (Show, Eq, Generic)
+
+makeLenses ''ATGNode
+makeLenses ''ATGEdge
+makeLenses ''ATG
