+++ core/src/Dscp/Core/Foundation/Educator/Certificate/Language.hs
@@ -2,4 +2,4 @@
 module Dscp.Core.Foundation.Educator.Certificate.Language where
 
 data Language = EN | RU
-    deriving (Show, Eq, Ord, Generic)
+    deriving (Show, Eq, Ord, Enum, Bounded, Generic)
